+++ app/models/bahanbeli.java
@@ -9,8 +9,8 @@ import play.db.jpa.Model;
 @Entity
 public class bahanbeli extends Model{
 	public String Nama_bahan ;
-	public long Stock;
-	public long Harga_Persatuan;
+	public double Stock;
+	public double Harga_Persatuan;
 	public Date Tanggal_Beli;
 	@ManyToOne
 	public satuan satuan;
