+++ app/models/bahanpakai.java
@@ -11,7 +11,7 @@ import play.db.jpa.Model;
 public class bahanpakai extends Model {
 	@ManyToOne
 	public bahan Nama_Bahan;
-	public long Stock;
+	public double Stock;
 	public Date Tanggal_Pakai;
 	public String oleh;
 }
